STPK-1429 Initial version of iio-sensors-hal component
This alternative sensors HAL should eventually be usable with all
sensors exposed using the iio interface. In the current form, it
has only been tested on a European T100 computer, which uses an
Invensense 6500 sensor complex, and only the accelerometer has
been taken care of.

The major reason for writing this new HAL module is the need for
flexibility to handle cases such as several sensors being exposed
through a single iio device. On the T100 the ALS and magnetometer
expose a polling interface (no buffer, no trigger, unreadable
character device). This kind of sensor is supported too.

Issue: STPK-1429

Change-Id: Ib4ebfec4d21559c929f3bbb7b4f2da452a50bd52
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>