STPK-1429 Unblock poll loop when a sensor gets enabled
This solves an issue where we could be blocked in our
event processing loop with an infinite timeout, and a
poll-mode sensor gets enabled. In that scenario no
reading associated to this sensor gets reported up,
because the poll timeout does not get re-evaluated.

Also interrupt poll on enabling as the timeout currently
being used may now be stale.

Issue STPK-1429

Change-Id: I7913463d8580862e16d67113d58a7596afdb7dae
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>