STPK-1429 Fix Klocwork issue #88803
Wise is the one who knows his limits.

There's an assumption in the code that enumerated all
sensors though, and that their reported indices are
contiguous. That should not be a problem with platforms
we control (e.g. MAX_SENSORS can be adjusted accordingly
as well as recognized sensors types), but that may be
worth revisiting in the future.

Issue: STPK-1429

Change-Id: I25392d1c4bf4c4a308f3c3448262445522fbd3be
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>