STPK-1429 Move immediate values acquisition to transform.c
This routine is in charge of reading and processing sensor
values for poll-mode sensors. It makes more sense to have
it within transform.c than control.c, as it applies offsets
and scales. So far we don't need Intel Sensor Hub -style
processing for these, because none of the ISH sensors uses
a polling interface.

Issue: STPK-1429

Change-Id: Ib45b4f9d32452c965b5829d17d9a57a9c9680902
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>