STPK-966: Fix orientation of accelerometer z-axis
This patch fixes the orientation of the accelerometer z coordinate.
It was inverted and it was causing two CTS Verifier tests to fail.

When the tablet was placed on a flat surface with the screen facing the
ceiling, the z-data of the accelerometer was negative and it should be
positive. This patch fixes this issue.

ISSUE: STPK-966
Change-Id: I4997e606aab10f65aa36a97b3fc2472b02e90be1
Signed-off-by: Ivan Gomez Castellanos <ivan.gomez.castellanos@intel.com>