STPK-1429 Enable the AOSP orientation sensor on T100
Google provides a software virtual sensor for orientation
but it is only created if we have hardware support for an
accelerometer, a gyroscope and a compass, if we don't
provide our own rotation vector sensor, and if we expose
an orientation sensor (which then gets replaced by the AOSP
virtual sensor). As we can't touch the shared SensorService
code, simply create a dummy orientation sensor if the right
conditions are satisfied, so Android can replace it with its
own implementation instead.

The sensor itself reacts strangely to certain movements, but
I believe that this is indicative of a problem in one of the
hardware sensors processing routines, which will be adressed
in a forthcoming commit.

Issue: STPK-1429

Change-Id: I97cf81685b0f9fd8c66153e806e9af6617efe0bd
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>