STPK-1429 Limit firing rate for poll-mode sensors
Don't refresh values associated to poll-mode sensors, and don't
raise the flag that will cause an event to be propagated upwards,
until an amount of time consistent with the sensor sampling rate
has elapsed since an event was last reported in relation to this
sensor.

This takes care of scenarios such as a low refresh rate poll-mode
sensor and a high refresh rate device-driven sensor being both
enabled, and a lot of events (and sysfs reads) being generated
for the low refresh rate sensor without a good reason to do so.

Issue: STPK-1429

Change-Id: I2bac3a8efd8da79bc642945b085408c29072d8c2
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>