STPK-1821 Work around what seems to be an ISH firmware bug
The hardware sensors controlled through the Intel Sensor Hub
sometimes do not send events before an explicit disable/enable
cycle, with a link to the initial sampling rate.

The issue does not happen if the Intel Sensor Hub firmware is
flashed with an up to date version, but let's take this opportunity
to introduce a quirk property that we'll be able to control slightly
altered behaviors on a sensor basis.

Issue STPK-1821

Change-Id: I93dc36214b0a3f545e3ff5b87d2d4e7ff7daf7f0
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>