STPK-1429 Add optional per-channel scaling factor
This adds support for a scaling factor that can be expressed
using properties, for all sensors, on a channel basis. That
scaling factor is applied after all other transforms have been
applied, and is intended as a way to compensate for problems
such as an incorrect axis polarity for a given sensor.

The syntax is <usual property prefix>.<channel>.scale, e.g.
ro.iio.accel.y.scale = -1 to negate the sign of the y readings
for the accelerometer.

For sensors using a single channel - and only those - the channel
name is implicitly void and a syntax such as ro.iio.illuminance.
scale = 3 has to be used.

Issue: STPK-1429

Change-Id: I240a614c0dd22ddfd036116cd4d1cf30be46003e
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>