STPK-1429 Remove articifial delay between poll exits
I introduced a minimum 1 ms delay between poll operations
early in development as a way to limit CPU usage in case
of problems with a sensor. That has become a significant
source of jitter now that we control a bunch of sensors
from that loop, and never was a good way to handle
misbehaving sensors anyway (we yet have to devise one).

This commits removes this forced gap between the last
poll exit and the next poll entry.

Issue: STPK-1429

Change-Id: I1b59ffbd71cad608456b98a336089c9057874aee
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>