STPK-1429 Cosmetic changes in code dealing with scales
Minor cosmetic adjustments, the most important being that we
don't emit a channel scale trace anymore if there is no sysfs
file expressing a scale, channel specific or otherwise. That's
the case with the dummy orientation sensor we're inserting as
a placeholder for replacement by the AOSP orientation sensor.

Issue: STPK-1429

Change-Id: I994e6049b42fd2e3963705f6fb99990ff2ffb4a6
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>