STPK-1429: Add 'order' configuration property
This allows field reordering for sensors that are tilted
inside the chassis, as this appears to be the case for the
ECS Cruise 7 tablet. This field reordering is applied in
the finalization stage, thus after opt_scale has been applied.

Ex: ro.iio.accel.order = 102 would swap x and y fields for
the accelerometer. Only digits are tolerated, and we must have
as many of them as there are fields for this sensor.

Issue: STPK-1429

Change-Id: I1f5eb3de0413bfa786408c7d59235bd486d09200
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>