STPK-1429: Fix trigger type for proximity sensor
The proximity sensor only sends an event if its state changes.

Note the statement at the end of sensor_activate about at least
one event being sent after a sensor is enabled. This will ensure
that whatever middleware components controls the proximity sensor,
it's not going to mistakenly believe that it's stuck in near state.

Change-Id: I56a8bd1ca171cd879d69c03fafab83dde9c42e1e
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>