IRDA-2056: Refine the test we use to decide to apply gyro calibration
The terse flag should only be used with drivers that a) don't
expose a recognizable motion trigger (we currently look for the
any-motion substring), b) are not for sensor types that Android
defines as continuous, and c) are nonetheless on-change. As of
today the only remaining example of this that I can think of
are the XPS 12 ISH drivers, so I prefer not to use this flag
to decide whether or not to apply calibration. Checking whether
we're using the initial or motion trigger seems better, if not
fully satisfactory.

Change-Id: I24c6632497f758e1301da7927ec94bf17f6fe84d
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>