IRDA-2056: Minor rework to the trigger switching code
Isolate the test that we use to switch from the motion trigger to
the continous one (for accelerometer based games that are sensitive
to small motion thresholds) into a function. Also, specifically
deactivate this behaviour for drivers marked as terse: for those it
may make sense to initialize the motion trigger name field with the
same string as the initial trigger name field, so let's check for
the terse flag.

Change-Id: Ia666caed8f04345aca7c09d6a686107590ffc590
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>