IRDA-2056: Rename continuous quirk flag and fix the location where it's used
This flag is used to force use of the continuous trigger, typically
in cases where the motion trigger would otherwise be autoselected.

Also, there are cases, such as with the Invensense driver that we
have on the T100, where a single trigger will control several
sensors because they're associated to the same iio device. In that
case having a single sensor forced to continous mode will imply that
no sensor co-located on the same iio device can be in "only send
events whenever there's motion" mode.

Change-Id: I4d01d2d964b70dead3e67c54bdd92d7d938cbf23
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>
Reviewed-on: https://android.intel.com:443/289220