GMINL-2659: Keep recent events history for fusion-like processing
We may wish to correlate recent input from several active sensors.
In preparation to that, keep the last 100 events we got for the
entire sensors collection. This is intended as a lightweight record,
built and used from the main sensor sample retrieval thread exclusively
so we don't have to deal with performance-degrading synchronization
primitives. This data will enable multi-sensors filtering decisions.

Change-Id: Ied6559c7dd91e94bebbeb1bc60182d1db3103d8c
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>