Validate old cal_level
In case we play with the cal level setting for a platform
it's best to check that the old one is not in fact outdated.
For example we lowered it for the ecs so the check is to make
sure this takes effect in case of an OTA update for example.

Change-Id: I78cba3722edd15ca23dc5f8c75bd8f0eef74d50e
Signed-off-by: Adriana Reus <adriana.reus@intel.com>