GMINL-3234: Reduce magneto and gyro filter windows
Use 20 samples for the magnetometer moving average, rather than
the previous 30. I believe that reduces latency that the camera
is sensitive to.

Also reduce the median filter set to 5 samples, rather than 7,
in the goal of increasing the "freshness" of gyro readings, with
an increase of noise level though.

At some point in the future we should make these per-sensor params.

Change-Id: I772c1fb720b5a1ebd9ed3d376bb038a1d0fed615
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-3234
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>