Fix min & max sampling rate for on-change sensors
The BMG160 thermometer was adopting the gyro min sampling rate of
100, which is kind of wasteful in terms of acquisition thread
activity. Same for BMC150 therm (using 15.6 Hz min). The ALS was
probably causing more thread switches than necessary, even if in
the end no event was being reported to Android if the value was
the same as the last measured one.

So, say it's fine to use on change sensors as a frequency as low
as 0.1 Hz, and report 5 Hz the maximum rate we can handle.

Change-Id: I7bd4a8ef97bffd5a0deb17cf581286214ce8a1a4
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>