Add the TYPE_MAGNETIC_FIELD_UNCALIBRATED sensor
Add the uncalibrated magnetometer type.
Added a single function for checking and adding virtual sensors.
Moved the timestamp handling for polling sensors so that we get a
timestamp before finalize is called.
Also moved the event drop (for sensors not directly enabled)
from finalize to propagate_sensor_report because for polling
sensors we need to signal an fd when we have an event,
and this is made by looking at the finalize return value.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-2744
Change-Id: I12abd38329ef7f59b2b13427c47a235b54f75fb2
Signed-off-by: Adriana Reus <adriana.reus@intel.com>