IRDA-3484: Add support for filter property
As of today, filtering is only engaged if a sensor has the noisy
quirk. If it is, we use median filtering for gyroscopes and moving
average for magnetometers. Let's allow for a filter configuration
item that would force any of the available filters selectively.

filter=median
filter=average

will supersede the default choice. There's no reason to support
a none setting, as it's the default.

Change-Id: I7f85738e3b770d86efe9a709495c08ca36ba4024
Tracked-On: https://jira01.devtools.intel.com/browse/IRDA-3484
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>