iio-sensors: Add sensor mode parameter
Currently the difference between a polled and a triggered sensor
is done using sensor_info[s].is_polling. If this is 1, the sensor
is polled, otherwise it is triggered.

If we need to handle more sensor types (e.g. sensors based on iio
events), this mode of differentiation is no longer sufficient.

Add a mode field to indicate the usage mode of the sensor: polled
or triggered. This will be later extended with other types (e.g.
event).

Change-Id: I3a6914a87413e6c8eb662e44f21b91a681e0643a
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>
Signed-off-by: Daniel Baluta <daniel.baluta@intel.com>