iio-sensors: Add support for STEP_COUNTER sensor
The step counter is a sensor that will count using a hardware
register the number of steps taken by the user since the last
reboot while activated. More information on the step counter in
Android can be found in Google's documentation:
http://source.android.com/devices/sensors/composite_sensors.html#counter.

The iio interface for the step counter is available with commit
a88bfe785 ("iio: core: Introduce STEPS channel, ENABLE mask and INSTANCE
event") soon to be found in linux kernel 3.20. Also the interface is
documented under Documentation/ABI/testing/sysfs-bus-iio

This is how the interface should be used:
- to enable the step counter, we need to write 1 to in_steps_en
- the current number of steps can be read from in_steps_input
- to disable the step counter, we need to write 0 to in_steps_en

This patch adds support for the step counter in the HAL.
The main difference from the other sensors is that the data
returned by the step counter is uint64_t, while other sensors
return float. This patch adds support for uint64_t return values
and adds the STEP_COUNTER in the supported sensors list.

Change-Id: I21a1288a9bf5ea02ae12176d1853dcd71e714bc9
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>
Signed-off-by: Daniel Baluta <daniel.baluta@intel.com>