IRDA-3621: Fix a bug where accel and gyro get stuck at 200 Hz
This is happening on Malata because the MPU-6050 driver exposes
both the accel and gyro on the same iio device, with a single
frequency knob for both. With the current code the accel would
be bumped up to 200 Hz if the gyro was configured at 200 Hz, but
would not go back when a lower sampling rate was selected because
the accel would appear to require 200 Hz.

Change-Id: Ied0c1c94c278be180a22fd537f8e82f0e8f7adbd
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>