[REVERTME] Revert "Pass iio provided timestamps without further processing."
Interrupts (events) are missed by the bmc150-accel & bmg160 drivers because it takes
too long to read from i2c. This might be caused by having three sensors on the same bus.
Process the timestamps in the HAL in order to make the CTSVerifier tests pass.
This commit reverts 58eef71e4b886632a7d51a6602fe8ef43b779d08 and it should be
reverted once the kernel problem is fixed.

Change-Id: Ieddd1e30253b9a79089740ca6a36e49b51c07d19
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6031
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6032
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6033
Tracked-By: https://jira01.devtools.intel.com/browse/GMINL-4337
Signed-off-by: Constantin Musca <constantin.musca@intel.com>