Fix LOCAL_MODULE for auto-detected targets
On GMINL, the iio sensor HAL is named sensors-$(TARGET_BOARD_PLATFORM).
For coho targets with auto-detection enabled, this HAL name does not
match the existing sensor records since they are not platform
dependent.

As in IRDA, the iio sensor HAL module name will be iio-sensors-hal.so
if auto-detection is enabled.

Issue: GMINL-6376
Change-Id: I83e09e9422b3e38edf5485aec6622214850e8247
Signed-off-by: Thierry Escande <thierry.escande@linux.intel.com>