GMINL-7944: Introduce internal sensor types for ALS
Decouple the sensor type retrieved from the catalog from the one
that is exposed to Android. This allows us to tailor processing
to specific iio sensor types. Start with the ALS, which comes in
several variants that we will want to arbitrate (i.e. if a sensor
exposes both illuminance and intensity, we'll keep illuminance).

This could enable finer processing per iio type, such as intensity
to lux conversions.

The type field in the sensor array becomes the internal type. The
type field on the sensor_desc array is the one exposed to Android.

Change-Id: Ie3a001d389af50f88732f71a3b06afc752255f4c
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>