GMINL-7944: Add quirks for disabling data acquisition methods
The current LTR-301 ALS driver is not exposing any events, but
expose channels, so we try to use it in trigger mode, only to
fail because it's not exposing any trigger name and doesn't
accept the default one (at least on ECS 10 edu, since no data
ready line is allocated). So let's force it to polling for now.

Change-Id: I910e0e52ebba536b81b0071dc9a201d11b846b35
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>