Add support for using a hrtimer trigger.
Add support for creating and using a hrtimer trigger
when we do not find any other suitable trigger exposed.
This may happen if for example we lack the appropriate interrupt
line for a platform.

We specify a "hrtimer" quirk to signal we allow creating one
for opne particular sensor. WE may not want this by default because
there are sensor drivers that do not specifically expose a trigger but
work if the traditional "-dev%d" one is setup.

Change-Id: I0c8a1f6c55b5258330370dc1ac77f8f6dae598cb
Signed-off-by: Adriana Reus <adriana.reus@intel.com>