sens : add poll with params feature for time or event orientated polling
Allows passing time and event number as parameters to the poll
method so that the polling will gracefully stop after an event exceeds
the targeted value or the timestamp exceeds the time allocated.

ex: sens poll 1000000000 22
Will print the number of events in 1 second or 22 events

Prints time remaining for the time orientated poll functions.

Change-Id: I77df6d9a0b6d35c5250747c1d149577894961a8e
Signed-off-by: Razvan Stamate <razvan.c.stamate@intel.com>
Signed-off-by: Maria Constantin <maria.p.constantin@intel.com>