GMINL-8061: Disregard the 100 Hz floor rate for polling BMG160
We don't need to respect the list of supported sampling rates when
we are not relying on the trigger to send us samples.

Also, there is no relation between the minimum rate that we can
activate on a sensor and the minimum rate that we should support
according to the CDD. That specific one tells that the chip ought
to support at least N Hz if requested to do so, not that we should
drive it at this frequency always.

Change-Id: I3007e8c1a16765e04ccc2865c5187ec7cfbed26d
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>