GMINL-10383: Allow for prioritization of sensors
This adds support for a 'secondary' quirk that ensures that
such sensors are listed after normal ones. One possible usage
is to list SAR proximity sensors after ALS ones.

i.e. setprop ro.iio.prox.sx9500.quirks "secondary"

The quirk is also implied by the presence of a negative
value in the last slot of the orientation matrix in the
Device Tree, for proximity sensors

Change-Id: I4d0b83b3220abfe7155b193e15356bb4e9ce8b1f
Signed-off-by: Patrick Porlan <patrick.porlan@intel.com>