Determine maxRange and resolution from IIO attributes
This patch determines the Android maxRange and resolution sensor
attributes based on the IIO attributes:

resolution = scale (or 1 if no scale is defined)

maxRange = 2^realbits * resolution + offset

If scale or offset are not defined at the sensor level we pick them
from the first channel.

This should avoid the need to define Android properties in most
cases. However, one can still use Android properties to override the
automatically computed values.

Change-Id: I04536be1fe741d887103ab07fda7daf4b4408ddc
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-14246
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Adriana Reus <adriana.reus@intel.com>
Reviewed-on: https://android.intel.com:443/392075