flush: use a pipe to unlock the sensor_poll func
If the framework calls the flush method while *sensor_poll*
is waiting for events then the flush_complete event is delayed until
the sensor reports the first event.
The CTSVerifier Sensor test (testBatchAndFlush) is failing for SX9500
due to this delay. (No event is reported because SX9500's reporting mode
is On-change).
Use a pipe to generate an event when the flush method is executed so that
the flush_complete event can be reported to the framework.

Change-Id: Ie24a32f546f71330bdd0205e7b6f398855c2146c
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-15425
Signed-off-by: Constantin Musca <constantin.musca@intel.com>
Reviewed-on: https://android.intel.com:443/397566